Some Implementations and Refactorings
